pyproject.toml
[tool.poetry]
name = "network-perf"
version = "0.1.0"
description = ""
authors = ["Konstantin Spirin <konstmsu@gmail.com>"]
readme = "README.md"
packages = [{include = "network_perf"}]

[tool.poetry.dependencies]
python = "^3.11"
speedtest-cli = "^2.1.3"


[tool.poetry.group.dev.dependencies]
black = "^23.10.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
